delete test file
